Replace Inline CSS Variables - Replace the inline CSS variable with it's corresponding value - Make sure a conversion from background -> background-color is happening in order for `iterateCSSDeclarations` to understand it. - Resolves #3914
