+++ src/inject/dynamic-theme/stylesheet-modifier.ts
@@ -2,7 +2,7 @@ import {Theme} from '../../definitions';
 import {createAsyncTasksQueue} from '../utils/throttle';
 import {iterateCSSRules, iterateCSSDeclarations, replaceCSSVariables} from './css-rules';
 import {getModifiableCSSDeclaration, ModifiableCSSDeclaration, ModifiableCSSRule} from './modify-css';
-import {isCSSStyleSheetConstructorSupported} from '../../utils/platform';
+import {getTempCSSStyleSheet} from 'inject/utils/dom';
 
 const themeCacheKeys: (keyof Theme)[] = [
     'mode',
@@ -18,24 +18,6 @@ const themeCacheKeys: (keyof Theme)[] = [
 
 function getThemeKey(theme: Theme) {
     return themeCacheKeys.map((p) => `${p}:${theme[p]}`).join(';');
-}
-
-let tempStyle: CSSStyleSheet = null;
-
-function getTempCSSStyleSheet(): CSSStyleSheet {
-    if (tempStyle) {
-        return tempStyle;
-    }
-    if (isCSSStyleSheetConstructorSupported) {
-        tempStyle = new CSSStyleSheet();
-        return tempStyle;
-    } else {
-        const tempStyleElement = document.createElement('style');
-        document.head.append(tempStyleElement);
-        tempStyle = tempStyleElement.sheet;
-        document.head.removeChild(tempStyleElement);
-        return tempStyle;
-    }
 }
 
 const asyncQueue = createAsyncTasksQueue();
